requirements.txt
pydantic==1.10.9
Pillow
scipy
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchvision
torchaudio
xformers
transformers
accelerate
safetensors
invisible-watermark>=0.2.0
diffusers